Remove GPX when removing workout, and fix not null constraint
When we remove a workout, we should clean up the associated GPX file;
otherwise, you can not add it back later.

We also override the "NOT NULL" constraint by setting a different tag.
This seems to be a weird workaround, but it's only for the time it needs
to get removed entirely, anyway.

This should fix #33

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>