Add auto-import feature
Each user can configure a folder from which to automatically import
workout files. This folder is then checked every minute, and each file
older than one minute is added.

When this process fails, the file is moved to the "failed" subfolder;
otherwise the file is added to the "done" subfolder. The subfolders are
created when needed.

Fixes #40

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>