Convert user statistics to use the new function
The data structure in the statistics.go file was updated to no longer add user
statistics. The UserStatistics template in the user_statistics.html file was
also updated to reflect the changes in the data structure.

The changes ensure that user statistics are retrieved and displayed correctly in
the user statistics page.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>