Add Distance method to Workout type and update user_profile_calendar.html
The Distance method has been added to the Workout type in the database
package. This method calculates and returns the total distance of the workout.
Additionally, the user_profile_calendar.html file has been updated to display
the distance of each workout in a human-readable format along with the user's
preferred distance units.