Switch the timestamp detection in a GPX file
This takes the first time from the first point in the GPX file as the
time of the workout, instead of the timestamp from the metadata. The
timestamp from the metadata is then used as fallback, but may have been
altered by converters.

We also change the auto-detection algorithm, so the limit of running vs
walking is now 7 instead of 5 km/h.

Fixes #57

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>