Parse extention nodes and standardize extra metric names
This adds better support for Apple Health GPX files, and provides a
standardized way to store the extra metric names.

We also parse a few extra metrics, but don't have a way to show them for
now (if ever). Only a visual cue to show whether there is heading
information in the GPX points.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>