fix: support editing workouts using Imperial units
In order to correctly support Imperial units and locales, we:
- change the input field for duration into split fields for hours,
  minutes and seconds (bonus: we now allow seconds)
- correctly convert the input distance from the user's preferred unit
  - for this, we add and use more constants

There's now a new reset button on the edit form that resets the input
fields back to their initial values, and a close button to go back to
the view.

We no longer show a few workout details in the list view:
- average speed without pauze
- max speed

We update the workout view page, to hide the map if there is no map to
show (even if it's a location-based workout). This is currently
determined by checking if the map center is "0, 0".

We also update the cosmetic layout of the buttons, to be more consistent
and intuitive.

Fixes: #217

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>