fix: add function to get filename of workout
There was a key that stored the filename of a workout, which was no
longer used and therefore removed in the previous PR (#232). This broke
the code for viewing and editing a workout, which still referenced the
old key.

We fix this by adding a function, that dynamically generates a new
filename if there was none, and checking whether we have data in the
views.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>