feat: Add calorie calculation and display
Adds calorie burned calculation and display on workout details page. The
calories are estimated based on the activity, assuming a weight of 70 kg. This
gives the user a rough idea of how many calories they burned during the workout.
Also adds repetition frequency per minute display on workout details page. This
is useful for understanding the intensity of a repetition-based workout.

Fixes #260

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>