feat: Add map full screen toggle
Adds a button to the workout detail page that allows toggling the map between
full width and a smaller size, depending on the screen size.

The map element now has a class that allows for CSS-based resizing. The size is
determined by media queries. This button is accessible with the "expand" icon,
and it is enabled by default when the socials are disabled.

Fixes #259