Style(css): Increase default heatmap height
The default height of the heatmap was increased from 300px to 600px for better
visualization. The smaller height made it difficult to see details, especially
on smaller screens. This change improves the user experience by providing a
clearer view of the data. Additionally, unused CSS classes were removed to clean
up the codebase.