feat(database): Improve statistics query
This commit improves the statistics query to handle the `since` parameter more
gracefully.

The query only adds a date filter if the `since` parameter is not
"forever". This change ensures that the statistics are calculated correctly for
all time periods. Additionally, the translation for "forever" has been added to
support this new functionality. Finally, the `since` options in the UI have been
updated to allow selection of "forever".

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>