fix(app): Fix redirect URLs after workout creation
The previous implementation used incorrect redirect URLs after creating or
updating a workout. This commit updates the redirect URLs to correctly point to
the "workout-add" route after creating a workout, and to the newly created
workout's "workout-show" route after saving changes. This ensures users are
redirected to the appropriate page after interacting with workouts.
Additionally, it handles the case of saving a new workout without an ID,
ensuring the `Workout.Data` is properly associated with the newly created
workout.

Fixes #412

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>