refactor: translations
This commit refactors the translation files to improve maintainability and
consistency. It moves the date range strings to the `misc` section of each
translation file. This is done to provide a more organized structure and make it
easier to manage common strings across different languages. The
StatisticSinceOptions and StatisticPerOptions functions have been updated to use
the new translation keys from the `misc` section. This ensures that the date
range options displayed in the user interface are consistent with the updated
translation structure.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>