feat(release): Add release tooling
Adds `git-semver` and `make release` targets to simplify the release process.

The `git-semver` tool is used to automatically determine the next semantic
version based on the commit history. The new make targets allow for easy
creation of releases with a single command.

The `release` target will tag the current commit with the new version and
provide instruction to push the tag to the remote repository and create the
actual release.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>