refactor(views): Move markdown rendering to helper
Moves the markdown rendering logic from the `RouteSegment` and `Workout`
database models to a shared helper function in `views/helpers`.

This separates presentation logic from data models, improving the architecture.

The new `MarkdownToHTML` helper is now used for rendering notes on the workout
and route segment show pages. It is also used to render the new translation help
text added to the login page.