feat(stats): Allow users to customize displayed total statistics
Handle empty workout types by returning "none" for their string representation.
This provides a clear identifier for an unselected workout type, which is then
used in UI elements.

Add 'none' and 'auto-detect' translations for sports.

Display the first statistic tile based on the user's `Profile.TotalsShow`
preference instead of the query results. This should be the same as the results
from the query, except when the query returns no results (in which case, the
value from the profile makes more sense!)

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>