fix(fit): Base the name of a workout on the local time
A FIT file does not contain a field directly useful for the name of a
workout.

The timestamp in a FIT file is UTC; it also contains a local timestamp
(but the timezone is still set as begin UTC, which is wrong).

We only know the actual location (and thus timezone) after geomapping,
which is deferred (and therefore not yet available).

We now use the local timestamp without timezone, prefixed with the
workout type, as the name for the workout.

Fixes #635

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>