fix(database): Correct workout average update and map detail saving
Change the condition for updating workout averages for a workout. Previously,
averages were updated only when a workout did not have an associated file. This
was incorrect, as averages should only be calculated and updated when there is
actual file data (e.g., GPS data) present to average. This correction ensures
that `UpdateAverages` is called only when meaningful data exists.

Add a nil check before attempting to save `MapData.Details`. `m.Details` can be
nil in some cases, and attempting to save a nil pointer with GORM can lead to
errors or unexpected behavior. This change ensures that `db.Save(m.Details)` is
only executed if `m.Details` is a valid (non-nil) object, improving the
robustness of the `MapData` persistence.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>